dataEtraction/requirements.txt
google-cloud-bigquery
fastapi
uvicorn
google-auth